requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic-settings==2.6.1
python-dotenv==1.0.1
pymongo==4.5.0
motor==3.4.0
# mcp==0.0.1
sentence-transformers==4.1.0
qdrant-client==1.12.1
python-docx==1.1.2
python-pptx==1.0.2
PyMuPDF==1.24.9
requests==2.32.3
openai==1.68.2
langchain-openai==0.3.17
langchain-community==0.3.5
langchain-groq==0.2.1
langchain-qdrant==0.2.0
langchain-google-genai==2.1.4